refactor: ResultCheckerにおける計算対象の一覧を別ファイルに移動
